Switch IceInternal::Buffer underlying type to std::byte (#1928)
Closes #1825

+++ matlab/src/Util.cpp
@@ -583,7 +583,7 @@ IceMatlab::getStringList(mxArray* m, vector<string>& v)
 }
 
 mxArray*
-IceMatlab::createByteArray(const uint8_t* begin, const uint8_t* end)
+IceMatlab::createByteArray(const byte* begin, const byte* end)
 {
     mxArray* r = mxCreateUninitNumericMatrix(1, end - begin, mxUINT8_CLASS, mxREAL);
     memcpy(reinterpret_cast<uint8_t*>(mxGetData(r)), begin, end - begin);
@@ -591,13 +591,13 @@ IceMatlab::createByteArray(const uint8_t* begin, const uint8_t* end)
 }
 
 mxArray*
-IceMatlab::createByteList(const vector<uint8_t>& bytes)
+IceMatlab::createByteList(const vector<byte>& bytes)
 {
     auto r = mxCreateCellMatrix(1, static_cast<int>(bytes.size()));
     mwIndex i = 0;
     for (auto byte : bytes)
     {
-        mxSetCell(r, i++, createByte(byte));
+        mxSetCell(r, i++, createByte(static_cast<uint8_t>(byte)));
     }
     return r;
 }
